requirements.txt
adbnx-adapter==0.0.0.2.5.1
appnope==0.1.2
arangopipe==0.0.6.9.3
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
boilerpipe3==1.3
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
charade==1.0.3
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.2
fuzzywuzzy==0.18.0
geographiclib==1.50
geojson==2.5.0
geopy==2.1.0
google==3.0.0
google-api-client==3.14.159265359.post1
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
googlemaps==4.4.5
graphviz==0.16
html2text==2020.1.16
httplib2==0.19.1
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.0
JPype1==1.2.1
jsonschema==3.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
mysql-connector-python==8.0.18
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
nose==1.3.7
notebook==6.2.0
numpy==1.19.2
olefile==0.46
openpyxl==3.0.6
OSMPythonTools==0.3.0
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pexpect==4.8.0
pgeocode==0.3.0
pickleshare==0.7.5
Pillow==8.1.0
pip==20.3.3
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.15.8
ptyprocess==0.7.0
py==1.10.0
pyArango==1.3.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.6.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.3
pytest-sugar==0.9.4
python-arango==7.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-Levenshtein==0.12.2
pytz==2021.1
pyzmq==20.0.0
qtconsole==5.0.2
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
rope==0.18.0
rsa==4.7.2
scikit-learn==0.23.2
scipy==1.6.0
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
setuptools-scm==6.0.1
six==1.15.0
soupsieve==2.1
sshtunnel==0.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.3
waybackpy==2.4.3
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
widgetsnbextension==3.5.1
wrapt==1.11.2
xarray==0.17.0
xlrd==2.0.1
zipp==3.4.0
zope.interface==5.4.0
